Fix merge failures and fix missing compilation when IR phase is not emitted + fix TestIRMatching

+++ test/hotspot/jtreg/compiler/lib/ir_framework/driver/irmatching/irrule/constraint/CountsConstraint.java
@@ -37,14 +37,14 @@ import compiler.lib.ir_framework.shared.Comparison;
  * @see Counts
  */
 public class CountsConstraint extends Constraint {
-    private final Comparison<Long> comparison;
+    private final Comparison<Integer> comparison;
 
-    public CountsConstraint(String regex, Comparison<Long> comparison, int index) {
+    public CountsConstraint(String regex, Comparison<Integer> comparison, int index) {
         super(regex, index);
         this.comparison = comparison;
     }
 
-    public Comparison<Long> getComparison() {
+    public Comparison<Integer> getComparison() {
         return comparison;
     }
 }
